Move `Watcher` to dedicated module
